requirements.txt
fastapi
uvicorn
torch
python-multipart
transformers
sentence-transformers
faiss-cpu
numpy
requests
elevenlabs
python-dotenv
openai
speechrecognition
fitz
pydub
pypdf
PyMuPDF==1.26.3
pypdf2        
python-docx
python-pptx